Add novpn build variants for tunnel and split tunnel
Introduce build-tagged vpn implementations that remove TUN and
split-tunnel
behavior in `novpn` builds while preserving API compatibility. Add
`novpn` stubs for split tunnel, force SOCKS-only mode with required
address validation, and split inbound/split-tunnel option construction
into default vs novpn files. Move shared split-tunnel types/constants to
a common file and gate affected tests with build tags, adding novpn-
specific coverage.
