+++ vpn/socks_only_default.go
@@ -1,0 +1,5 @@
+//go:build !novpn
+
+package vpn
+
+func socksOnlyEnforced() bool { return false }
