+++ vpn/split_tunnel_test.go
@@ -1,3 +1,5 @@
+//go:build !novpn
+
 package vpn
 
 import (
