+++ vpn/split_tunnel_types.go
@@ -1,0 +1,68 @@
+package vpn
+
+import (
+	"fmt"
+	"strings"
+)
+
+// splitTunnelTag is the route rule-set tag the split-tunnel filter is bound to.
+// It is shared across builds so the novpn build can reference the same tag when
+// asserting the rule-set is absent.
+const splitTunnelTag = "split-tunnel"
+
+// Filter type identifiers accepted by SplitTunnel.AddItems/RemoveItems and exposed
+// through the backend and CLI. Defined here (rather than alongside the real
+// implementation) so the novpn build, which compiles an inert SplitTunnel, still
+// exports the same identifiers to callers.
+const (
+	TypeDomain           = "domain"
+	TypeDomainSuffix     = "domainSuffix"
+	TypeDomainKeyword    = "domainKeyword"
+	TypeDomainRegex      = "domainRegex"
+	TypeProcessName      = "processName"
+	TypeProcessPath      = "processPath"
+	TypeProcessPathRegex = "processPathRegex"
+	TypePackageName      = "packageName"
+)
+
+// SplitTunnelFilter is the set of domains, processes, and packages a user has
+// configured to bypass (or be confined to) the tunnel.
+type SplitTunnelFilter struct {
+	Domain           []string
+	DomainSuffix     []string
+	DomainKeyword    []string
+	DomainRegex      []string
+	ProcessName      []string
+	ProcessPath      []string
+	ProcessPathRegex []string
+	PackageName      []string
+}
+
+func (f SplitTunnelFilter) String() string {
+	var str []string
+	if len(f.Domain) > 0 {
+		str = append(str, fmt.Sprintf("domain: %v", f.Domain))
+	}
+	if len(f.DomainSuffix) > 0 {
+		str = append(str, fmt.Sprintf("domainSuffix: %v", f.DomainSuffix))
+	}
+	if len(f.DomainKeyword) > 0 {
+		str = append(str, fmt.Sprintf("domainKeyword: %v", f.DomainKeyword))
+	}
+	if len(f.DomainRegex) > 0 {
+		str = append(str, fmt.Sprintf("domainRegex: %v", f.DomainRegex))
+	}
+	if len(f.ProcessName) > 0 {
+		str = append(str, fmt.Sprintf("processName: %v", f.ProcessName))
+	}
+	if len(f.ProcessPath) > 0 {
+		str = append(str, fmt.Sprintf("processPath: %v", f.ProcessPath))
+	}
+	if len(f.ProcessPathRegex) > 0 {
+		str = append(str, fmt.Sprintf("processPathRegex: %v", f.ProcessPathRegex))
+	}
+	if len(f.PackageName) > 0 {
+		str = append(str, fmt.Sprintf("packageName: %v", f.PackageName))
+	}
+	return "{" + strings.Join(str, ", ") + "}"
+}
